prov/efa: Implement eager protocol send completion callback
Move the eager RTM send completion handling from efa_rdm_pke_rtm.c into
efa_rdm_proto_eager.c.

Replace the old eager case in efa_rdm_pke_handle_send_completion with an
assertion to catch any packets that incorrectly bypass the new callback.
